Add connectTimeout LdapConnection field with default value of 5000 ms.
This is the timeout value to use when creating new TCP connections.
In general this is most obvious when attempting to connect to the LDAP
server for the first time. If it is not reachable one would normally want
the connection attempt to end ASAP.

For more info the node-ldapjs docs at:
https://github.com/ldapjs/node-ldapjs/blob/8ffd0bc9c149088a10ec4c1ec6a18450f76ad05d/docs/client.md#create-a-client